crud image portfolio

+++ src/components/dashboard/portfolio/index.jsx
@@ -3,11 +3,27 @@ import ButtonComponents from "@/components/ButtonComponents";
 import Link from "next/link";
 import { useEffect, useState } from "react";
 import CustomAlert from "../../AlertComponents";
+import { useRouter } from "next/router";
 
 const DashboardPortfolio = () => {
   const [porto, setPorto] = useState([]);
-  const [deleteSuccess, setDeleteSuccess] = useState(false);
+  let [deleteSuccess, setDeleteSuccess] = useState(false);
+  let [createSuccess, setCreateSuccess] = useState(false);
+  const router = useRouter();
+
   useEffect(() => {
+    setDeleteSuccess(router.query.deleteSuccess === "true");
+    setCreateSuccess(router.query.createSuccess === "true");
+    if (
+      router.query.createSuccess === "true" ||
+      router.query.deleteSuccess === "true"
+    ) {
+      const { pathname, query } = router;
+      const newQuery = { ...query };
+      delete newQuery.createSuccess;
+      delete newQuery.deleteSuccess;
+      router.replace({ pathname, query: newQuery });
+    }
     const fetchData = async () => {
       try {
         const response = await fetch("/api/portfolio");
@@ -26,12 +42,22 @@ const DashboardPortfolio = () => {
   }, []);
   const handleDelete = async (id) => {
     try {
+      const deleteImageResponse = await fetch(
+        `/api/portfolio/image/edit/${id}`,
+        {
+          method: "DELETE",
+        }
+      );
+      if (!deleteImageResponse.ok) {
+        throw new Error("Failed to delete image");
+      }
       const response = await fetch(`/api/portfolio/${id}`, {
         method: "DELETE",
       });
-      setDeleteSuccess(true);
       if (!response.ok) {
         throw new Error("Failed to delete data");
+      } else {
+        setDeleteSuccess(true);
       }
 
       // Update the state to reflect the changes
@@ -46,12 +72,22 @@ const DashboardPortfolio = () => {
         <Link href={"/dashboard/portfolio/create"} className="p-3">
           <ButtonComponents textButton="Add New" />
         </Link>
+        <Link href={"/dashboard/portfolio/addcategory"} className="p-3">
+          <button className="btn btn-success"> + Add New Category</button>
+        </Link>
         <div className="container-fluid mt-3">
           {deleteSuccess && (
             <CustomAlert
               variant="danger"
               message="Data deleted successfully!"
-              onClose={() => setUpdateSuccess(false)}
+              onClose={() => setDeleteSuccess(false)}
+            />
+          )}
+          {createSuccess && (
+            <CustomAlert
+              variant="success"
+              message="Data created successfully!"
+              onClose={() => setCreateSuccess(false)}
             />
           )}
           <div className="row">
