+++ src/components/dashboard/portfolio/show.jsx
@@ -10,39 +10,72 @@ const ShowPortfolio = () => {
   const { id } = router.query;
   const [portfolio, setPortfolio] = useState([]);
   const [deleteSuccess, setDeleteSuccess] = useState(false);
-  useEffect(() => {
-    const fetchDataById = async () => {
-      try {
-        if (!id) {
-          return;
-        }
-        const rute = `/api/portfolio/${id}`;
-        const response = await fetch(rute);
-        if (!response.ok) {
-          throw new Error("Failed to fetch data");
-        }
 
-        const data = await response.json();
-        setPortfolio(data);
-      } catch (error) {
-        console.error("Error fetching portfolio data:", error);
-      }
+  let isMounted = true;
+
+  useEffect(() => {
+    // Set isMounted to false when the component is unmounted
+    return () => {
+      isMounted = false;
     };
-    fetchDataById();
-  }, [id, portfolio]);
+  }, []);
+
+  useEffect(() => {
+    // Hanya fetch data jika ID ada
+    if (id) {
+      fetchDataById();
+    }
+  }, [id]);
+
+  const fetchDataById = async () => {
+    try {
+      const rute = `/api/portfolio/${id}`;
+      const response = await fetch(rute);
+      if (!response.ok) {
+        throw new Error("Failed to fetch data");
+      }
+
+      const data = await response.json();
+      setPortfolio(data);
+    } catch (error) {
+      console.error("Error fetching portfolio data:", error);
+    }
+  };
+
   const handleDelete = async (id) => {
     try {
-      const response = await fetch(`/api/portfolio/${id}`, {
-        method: "DELETE",
-      });
-      setDeleteSuccess(true);
-      router.push("/dashboard/portfolio");
-      if (!response.ok) {
-        throw new Error("Failed to delete data");
+      const imageInfoResponse = await fetch(
+        `/api/portfolio/image/${portfolio.filename}`
+      );
+      if (imageInfoResponse.ok) {
+        const deleteImageResponse = await fetch(
+          `/api/portfolio/image/edit/${id}`,
+          {
+            method: "DELETE",
+          }
+        );
+        if (!deleteImageResponse.ok) {
+          throw new Error("Failed to delete image");
+        }
       }
 
-      // Update the state to reflect the changes
-      setPorto((prevPorto) => prevPorto.filter((item) => item.id !== id));
+      const deletePortfolioResponse = await fetch(`/api/portfolio/${id}`, {
+        method: "DELETE",
+      });
+
+      if (!deletePortfolioResponse.ok) {
+        throw new Error("Failed to delete portfolio data");
+      }
+      setDeleteSuccess(true);
+      if (isMounted) {
+        setPortfolio([]);
+      }
+
+      // Redirect ke /dashboard/portfolio dengan parameter query deleteSuccess=true
+      router.push({
+        pathname: "/dashboard/portfolio",
+        query: { deleteSuccess: true },
+      });
     } catch (error) {
       console.error("Error deleting portfolio data:", error);
     }
@@ -71,7 +104,13 @@ const ShowPortfolio = () => {
       </div>
       <div className="d-flex justify-content-center">
         <Card className="mt-5 card-porto-admin">
-          <Card.Img variant="top" src="/images/img-lanscape.png" width="100%" />
+          {portfolio.path && (
+            <Card.Img
+              variant="top"
+              src={`/api/portfolio/image/${portfolio.filename}`}
+              width="100%"
+            />
+          )}
           <Card.Body className="card-body-porto">
             <div className="card-text">
               <div className="title-portfolio">{portfolio.title}</div>
